// src/model/side-type.ts
/**
 * Represents the possible sides of a rectangular shaped node where a path can be attached.
 */
export enum SideType {
    RIGHT = 'RIGHT',
    BOTTOMRIGHT = 'BOTTOMRIGHT',
    BOTTOM = 'BOTTOM',
    BOTTOMLEFT = 'BOTTOMLEFT',
    LEFT = 'LEFT',
    TOPLEFT = 'TOPLEFT',
    TOP = 'TOP',
    TOPRIGHT = 'TOPRIGHT'
}